Product Selection, page, more (IMPORTANT)
Now when u click on a product it takes you to another page with the product shown, with its cost and name. Javascript files added. This required the products to bee changed to a link, inorder to refernec the product for later. NEW CSS file for the new product html (For displaying the html), I did this to make it easier to format for me, because Im getting lost in the clutter. Tried to make it easier to read by adding headings. Tidied up some areas.  And for the Eyeglasses page, in figma I didnt make a scroller for it, so its just boxes on a page, if u check the design, so that needs to be formmatted again, i did a bit though.

Sorry, if I'm doing to much.